Make the forked child rearrange itself
All the functions rearrange themselves immediately
after fork(). The original code is marked READ (for got)
and NOEXEC (for other code).

The Makefile change is required so that callq are not
relative to the position of the code in the object file.
This is required because we are moving the function
code around in memory before executing.